+++ src/main.rs
@@ -1,133 +1,5 @@
-use serde::{Deserialize, Serialize};
-use serde_json::json;
-use std::error;
-
-#[derive(Debug, Serialize, Deserialize)]
-struct Label {
-    id: i64,
-    name: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct Issue {
-    id: i64,
-    title: String,
-    url: String,
-    state: String,
-    body: Option<String>,
-    labels: Option<Vec<Label>>,
-}
-
-#[derive(Debug, Clone)]
-struct GetIssueError {
-    message: String,
-}
-
-// implement for GetIssueError
-impl std::fmt::Display for GetIssueError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self.message)
-    }
-}
-
-impl error::Error for GetIssueError {}
-
-fn get_github_personal_access_token() -> String {
-    // get token from environment variable
-    std::env::var("GITHUB_PERSONAL_ACCESS_TOKEN").unwrap()
-}
-
-fn get_slack_webhook_url_from_env() -> String {
-    // get token from environment variable
-    std::env::var("SLACK_WEBHOOK_URL").unwrap()
-}
-
-async fn get_my_issues() -> Result<Vec<Issue>, GetIssueError> {
-    let token = get_github_personal_access_token();
-
-    let client = reqwest::Client::new();
-    let res = client
-        .get("https://api.github.com/issues?filter=assigned&state=open&labels=Priority: High")
-        .header("User-Agent", "reqwest")
-        .header("Accept", "application/vnd.github+json")
-        .header("Authorization", format!("Bearer {}", token))
-        .send()
-        .await;
-    // if res is Ok, print "OK", otherwise print "Error"
-    let issues = match res {
-        Ok(res) => {
-            // if status is not 200, return error
-            if res.status() != 200 {
-                return Err(GetIssueError {
-                    message: format!("status code is not 200: {}", res.status()),
-                });
-            }
-
-            println!("Fetch Issues OK");
-            let parsed_issues = res.json::<Vec<Issue>>().await;
-            match parsed_issues {
-                Ok(issues) => issues,
-                Err(e) => {
-                    let error_message = format!("Parse Issues Error: {}", e);
-                    println!("{}", error_message);
-                    return Err(GetIssueError {
-                        message: error_message,
-                    });
-                }
-            }
-        }
-        Err(err) => {
-            let error_message = format!("Fetch Issues Error: {}", err);
-            println!("{}", error_message);
-            return Err(GetIssueError {
-                message: error_message,
-            });
-        }
-    };
-    Ok(issues)
-}
-
-async fn notify_by_slack(text: String) {
-    let webhook_url = get_slack_webhook_url_from_env();
-    let client = reqwest::Client::new();
-    let res = client
-        .post(webhook_url)
-        .json(&json!({ "text": text }))
-        .send()
-        .await;
-    match res {
-        Ok(_res) => {
-            println!("Notify by Slack OK");
-        }
-        Err(err) => {
-            println!("Notify by Slack Error: {}", err);
-        }
-    }
-}
-
-fn create_payload_for_slack(issues: Result<Vec<Issue>, GetIssueError>) -> String {
-    let mut payload = String::new();
-
-    match issues {
-        Ok(issues) => {
-            payload.push_str("@channel\n優先度が高いタスク一覧\n");
-            if issues.is_empty() {
-                payload = "なし".to_string();
-            } else {
-                for issue in issues {
-                    let issue_url = issue.url;
-                    let issue_title = issue.title;
-                    payload.push_str(&format!("- <{}|{}>\n", issue_url, issue_title));
-                }
-            }
-        }
-        Err(e) => {
-            payload = e.message;
-        }
-    }
-
-    payload
-}
+use my_github_notification::github::get_my_issues;
+use my_github_notification::slack::{create_payload_for_slack, notify_by_slack};
 
 #[tokio::main]
 async fn main() {
